Rename function tangentEnabled to isTangentEnabled Correct all the warning with VS.NET

+++ cal3d/src/cal3d/animation_cycle.cpp
@@ -183,7 +183,7 @@ bool CalAnimationCycle::update(float deltaTime)
     m_time += deltaTime * m_timeFactor;
     if(m_time >= m_pCoreAnimation->getDuration())
     {
-      m_time = fmod(m_time, m_pCoreAnimation->getDuration());
+      m_time = (float) fmod(m_time, m_pCoreAnimation->getDuration());
     }
   }
